New sql_wait_secs option

+++ sh/postgres_functions.sh
@@ -38,6 +38,8 @@
 # fc_db_pre_exec_call
 # fc_db_sql_transform
 
+########################
+
 # printf %s\\n "$-"
 bin_check_exit psql
 bin_check_exit mkfifo
@@ -48,9 +50,11 @@ fc_db_startup_connection ()
   [ -f "${v_database_fifo_file}" -o -p "${v_database_fifo_file}" ] && rm -f "${v_database_fifo_file}"
   mkfifo "${v_database_fifo_file}"
   exec 3<>"${v_database_fifo_file}"
+  fc_def_output_file v_database_in_file 'database_input.log'
   fc_def_output_file v_database_out_file 'database_output.log'
   fc_def_output_file v_database_err_file 'database_error.log'
   cat <&3 | psql --file - ${moat370_sw_db_conn_params} > "${v_database_out_file}" 2> "${v_database_out_file}" &
+  # cat <&3 | psql ${moat370_sw_db_conn_params} > "${v_database_out_file}" 2> "${v_database_out_file}" &
   v_db_client_pid=$!
 }
 
@@ -60,39 +64,22 @@ fc_db_end_connection ()
   [ -p "${v_database_fifo_file}" ] && rm -f "${v_database_fifo_file}"
   fc_seq_output_file v_database_out_file
   fc_zip_file "${moat370_zip_filename}" "${v_database_out_file}"
-  kill_db_connection
+  db_connection_kill
 }
 
 fc_db_check_connection ()
 {
-  local v_loop_limit v_sleep_time v_total_sleep
-  v_loop_limit=15
-  v_sleep_time=1
-  v_total_sleep=0
-
   fc_run_query '\pset tuples_only on'
   fc_run_query "SELECT CONCAT('I_AM_CONNECTED_',COUNT(*));"
   fc_run_query '\pset tuples_only off'
 
-  set +x
-
-  while :
-  do
-    if grep -q 'I_AM_CONNECTED_1' "${v_database_out_file}"
-    then
-      break
-    fi
-    sleep ${v_sleep_time}
-    v_total_sleep=$(do_calc "v_total_sleep+1")
-    if [ ${v_total_sleep} -gt ${v_loop_limit} ]
-    then
-      cat "${v_database_out_file}"
-      exit_error "Unable to connect on database.. time limit exceeded."
-    fi
-  done
-
-  fc_enable_set_x
-
+  moat370_check_connection_status=false
+  if ! fc_wait_string "${v_database_out_file}" I_AM_CONNECTED_1 10 2> /dev/null
+  then
+    cat "${v_database_out_file}"
+    exit_error "Unable to connect on database.. time limit exceeded."
+  fi
+  moat370_check_connection_status=true
 }
 
 fc_db_begin_code ()
@@ -124,7 +111,10 @@ fc_db_define_module ()
 
 fc_db_reset_options ()
 {
-  fc_db_run_file "${moat370_fdr}"/database/postgres/postgres_reset.sql
+  if [ -f "${moat370_fdr}"/database/postgres/postgres_reset.sql ]
+  then
+    fc_db_run_file "${moat370_fdr}"/database/postgres/postgres_reset.sql
+  fi
 }
 
 fc_db_pre_section_call ()
@@ -134,6 +124,8 @@ fc_db_pre_section_call ()
 
 fc_db_create_csv ()
 {
+  local v_ret
+
   local v_in_file="$1"
   local v_out_csv="$2"
 
@@ -148,23 +140,29 @@ fc_db_create_csv ()
   fc_replace_file_variable "${v_in_tmp}" '%%query_file%%' "${v_in_file}"
   rm -f "${v_database_err_file}"
   fc_run_query "\include ${v_in_tmp}"
-  fc_check_executed
+  fc_check_executed ${sql_wait_secs} && v_ret=$? || v_ret=$?
+
   rm -f "${v_in_tmp}"
   unset v_in_tmp
 
-  fc_convert_html_to_csv "${v_out_html}" "${v_out_csv}"
-  rm -f "${v_out_html}"
-  unset v_out_html
-
-  # This is not required as html characters are already converted.
-  # fc_escape_markup_characters "${v_out_csv}" > "${v_out_csv}".tmp
-  # mv "${v_out_csv}".tmp "${v_out_csv}"
-
-  $cmd_sed -e ':a' -e '/./,$!d;/^\n*$/{$d;N;};/\n$/ba' "${v_out_csv}" > "${v_out_csv}".tmp
-  mv "${v_out_csv}".tmp "${v_out_csv}"
-
-  #fc_load_variable MOAT370_PREV_SQL_ID moat370_prev_sql_id
-  #fc_load_variable MOAT370_PREV_CHILD_NUMBER moat370_prev_child_number
+  if [ $v_ret -eq 0 ]
+  then
+    fc_convert_html_to_csv "${v_out_html}" "${v_out_csv}"
+    rm -f "${v_out_html}"
+    unset v_out_html
+  
+    # This is not required as html characters are already converted.
+    # fc_escape_markup_characters "${v_out_csv}" > "${v_out_csv}".tmp
+    # mv "${v_out_csv}".tmp "${v_out_csv}"
+  
+    $cmd_sed -e ':a' -e '/./,$!d;/^\n*$/{$d;N;};/\n$/ba' "${v_out_csv}" > "${v_out_csv}".tmp
+    mv "${v_out_csv}".tmp "${v_out_csv}"
+  
+    #fc_load_variable MOAT370_PREV_SQL_ID moat370_prev_sql_id
+    #fc_load_variable MOAT370_PREV_CHILD_NUMBER moat370_prev_child_number
+  else
+    fc_echo_screen_log "${sql_wait_secs}s timeout reached."
+  fi
 
   fc_db_reset_options
 }
@@ -182,16 +180,22 @@ fc_db_create_raw ()
   fc_replace_file_variable "${v_in_tmp}" '%%query_file%%' "${v_in_file}"
   rm -f "${v_database_err_file}"
   fc_run_query "\include ${v_in_tmp}"
-  fc_check_executed
+  fc_check_executed ${sql_wait_secs} && v_ret=$? || v_ret=$?
+
   rm -f "${v_in_tmp}"
   unset v_in_tmp
+
+  if [ $v_ret -ne 0 ]
+  then
+    fc_echo_screen_log "${sql_wait_secs}s timeout reached."
+  fi
 
   fc_db_reset_options
 }
 
 fc_db_table_description ()
 {
-  local v_output_file
+  local v_output_file v_ret
   v_output_file="$1"
 
   fc_def_output_file v_in_tmp 'fc_db_table_description.tmp'
@@ -200,12 +204,13 @@ fc_db_table_description ()
   fc_replace_file_variable "${v_in_tmp}" '%%file_name%%' "${v_output_file}.desc"
   fc_replace_file_variable "${v_in_tmp}" '%%table_name%%' "${main_table}"
   fc_run_query "\include ${v_in_tmp}"
-  fc_check_executed
+  fc_check_executed 5
+  v_ret=$?
   rm -f "${v_in_tmp}"
   unset v_in_tmp
 
   # Remove desc file.
-  cat "${v_output_file}.desc" >> "${v_output_file}"
+  [ ${v_ret} -eq 0 ] && cat "${v_output_file}.desc" >> "${v_output_file}"
   rm -f "${v_output_file}.desc"
 }
 
@@ -227,9 +232,6 @@ fc_db_check_file_sql_error ()
 fc_db_enable_trace ()
 {
   true
-  # fc_run_query "ALTER SESSION SET MAX_DUMP_FILE_SIZE='1G';"
-  # fc_run_query "ALTER SESSION SET TRACEFILE_IDENTIFIER=\"${moat370_tracefile_identifier}\";"
-  # fc_run_query "ALTER SESSION SET EVENTS '10046 TRACE NAME CONTEXT FOREVER, LEVEL ${sql_trace_level}';"
 }
 
 fc_db_pre_exec_call ()
@@ -250,10 +252,17 @@ fc_db_sql_transform ()
 fc_run_query ()
 {
   echo "${1}" > "${v_database_fifo_file}"
+  echo_time "${1}" >> "${v_database_in_file}"
 }
 
 fc_check_executed ()
 {
+  # Optional parameter for how long (seconds) will wait.
+  set +u
+  local v_wait_limit="$1"
+  fc_enable_set_u
+
+  local v_ret
   # Check if database FIFO file reached the end.
   fc_def_output_file v_fc_db_run_file_output 'run_file.out'
   rm -f "${v_fc_db_run_file_output}"
@@ -261,15 +270,26 @@ fc_check_executed ()
   fc_run_query "\out ${v_fc_db_run_file_output}"
   fc_run_query "\qecho FINISHED"
   fc_run_query "\out"
-  fc_wait_string "${v_fc_db_run_file_output}" FINISHED
+
+  fc_wait_string "${v_fc_db_run_file_output}" FINISHED ${v_wait_limit} 2> /dev/null && v_ret=$? || v_ret=$?
+
+  # Kill current database oparation to move on.
+  if [ $v_ret -ne 0 ]
+  then
+    echo_error "Query timed out. Restarting the connection..."
+    db_connection_sigint
+    db_connection_kill
+    fc_connection_flow
+  fi
 
   rm -f "${v_fc_db_run_file_output}"
+  return ${v_ret}
 }
 
 fc_load_variable ()
 {
   # This code will convert USER DEFINED variables into SHELL variables.
-  local v_load_variable_file v_load_variable_name v_output v_load_variable_result
+  local v_load_variable_file v_load_variable_name v_output v_load_variable_result v_ret
   v_load_variable_name="$1"
 
   set +u
@@ -287,30 +307,24 @@ fc_load_variable ()
     fc_replace_file_variable "${v_temp_variable_file}" '%%var_name%%' "${v_load_variable_name}"
     fc_replace_file_variable "${v_temp_variable_file}" '%%file_name%%' "${v_load_variable_file}"
     fc_run_query "\include ${v_temp_variable_file}"
-    fc_check_executed
+    fc_check_executed 5
+    v_ret=$?
 
-    v_load_variable_result=$(cat "${v_load_variable_file}")
-
-    if [ "${v_load_variable_result}" = ":${v_load_variable_name}" ]
+    if [ $v_ret -eq 0 ]
     then
-      echo_error "Undefined variable ${v_load_variable_name}."
-      return
+      v_load_variable_result=$(cat "${v_load_variable_file}")
+      if [ "${v_load_variable_result}" = ":${v_load_variable_name}" ]
+      then
+        echo_error "Undefined variable ${v_load_variable_name}."
+        return
+      fi
+      eval ${v_load_variable_name}=\${v_load_variable_result}
     fi
-
-    eval ${v_load_variable_name}=\${v_load_variable_result}
-
-  # TODO
-  # else
-  #   cp "${moat370_fdr}/database/postgres/postgres_load_variable_all.sql" "${v_temp_variable_file}"
-  #   fc_replace_file_variable "${v_temp_variable_file}" '%%file_name%%' "${v_load_variable_file}"
-  #   fc_run_query "\include ${v_temp_variable_file}"
-  #   fc_check_executed
   fi
 
   # source "${v_load_variable_file}"
   rm -f "${v_temp_variable_file}"
   rm -f "${v_load_variable_file}"
-
 }
 
 fc_convert_html_to_csv ()
@@ -342,16 +356,4 @@ fc_convert_html_to_csv ()
   tr '\f' '\n' | \
   # Remove any other HTML line and empty lines.
   $cmd_sed '/^</d' > "${v_out_file}"
-}
-
-fc_replace_file_variable ()
-{
-  local v_in_file="$1"
-  local v_in_attr=$(ere_quote "$2")
-  local v_in_repl=$(ere_quote "$3")
-  local v_out_file="${v_in_file}.tmp"
-
-  sed "s|${v_in_attr}|${v_in_repl}|g" "${v_in_file}" > "${v_out_file}"
-
-  mv "${v_out_file}" "${v_in_file}"
 }
